requirements.txt
# Core ML
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3

# Experiment Tracking & Registry
mlflow==2.10.2

# Data Versioning
dvc>=3.0.0

# Data Validation
pydantic==2.5.3
great-expectations==0.18.15

# API & Serving
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Monitoring
evidently==0.4.20
scipy>=1.12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
hydra-core==1.3.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Linting
ruff==0.1.14

# Gradio (for Hugging Face Spaces)
gradio>=4.0.0
huggingface_hub>=0.20.0
